made design changes to carousel, landing page, movieplayer
