+++ src/components/cards/cards.css
@@ -15,11 +15,11 @@
     margin-bottom: 30px;
 
 }
-.movie-poster img {
+/* .movie-poster img {
     width: 100%;
     height: 376px;
     border-radius:10px;
-}
+} */
 .movie-poster {
     position: relative;
 }
@@ -33,12 +33,20 @@
     background-color: rgba(0, 0, 0, 0.7);
     display: none;
 }
-.movie-poster-overlay img {
-    width: 200px;
-    height: 200px;
+.movie-poster-overlay button {
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    border: none;
+    background-color: rgba(0, 0, 0, 0.8);
+    color: #FFF;
+    font-size: 30px;
     position: absolute;
-    top: 80px;
-    left: 25px;
+    display:flex;
+    align-items: center;
+    justify-content: center;
+    top: 180px;
+    left: 100px;
 
 }
 .movie-poster:hover .movie-poster-overlay {
