+++ src/components/tabs/carousel.js
@@ -3,15 +3,11 @@ import { CgSpinnerTwo } from 'react-icons/cg';
 import axios from "axios"
 import "./carousel.css";
 import { MovieModal } from "../../components";
-import AOS from "aos";
-import 'aos/dist/aos.css';
-import {Playicon} from "../../assets";
-import { FiPlay } from "react-icons/fi";
 
 
 const CarouselPage = () => {
 
-    AOS.init();
+
 const [movies, setMovies] = useState([]);
 const [isLoading, setIsLoading] = useState(true);
 
@@ -44,13 +40,11 @@ const toggleMovieModal = (e, movie) => {
             </h2>
           <section className="slider-container">
                 {movies.map(movie => (
-                    <div key={movie._id} className="slider-content-container" data-aos="fade-up" data-aos-duration="1000">
-                        <section className="movie-poster">
-                        <img src={movie.coverpics_url} alt="poster" />
-                        <span className="movie-poster-overlay" >
-                            <button onClick={(e) => toggleMovieModal(e, movie)}><FiPlay/></button>
+                    <div key={movie._id} className="slider-content-container">
+                        <img src={movie.coverpics_url} alt="coverpic"/>
+                        <span className="movie-poster-overlay">
+                        <button onClick={(e) => toggleMovieModal(e, movie)}>click me</button>
                         </span>
-                    </section>
                         <div className="carousel-movie-details-container">
                             <h5>
                                 {movie.title}
@@ -58,7 +52,7 @@ const toggleMovieModal = (e, movie) => {
                             <ul className="carousel-movie-details">
                                 <li>{movie.release_date}</li>
                                 <li>{movie.category}</li>
-                                <li>{movie.length}</li>
+                                <li>2hr 3mins</li>
                             </ul>
                         </div>
                     </div>
@@ -108,13 +102,11 @@ const toggleMovieModal = (e, movie) => {
             </h2>
           <section className="recent-upload-container">
                 {movies.map(movie => (
-                    <div key={movie._id} className="recent-upload-content-container" data-aos="fade-up" data-aos-duration="1000">
-                       <section className="movie-poster">
-                        <img src={movie.coverpics_url} alt="poster" />
-                        <span className="movie-poster-overlay" >
-                            <button onClick={(e) => toggleMovieModal(e, movie)}><FiPlay/></button> 
+                    <div key={movie._id} className="recent-upload-content-container">
+                        <img src={movie.coverpics_url} alt="coverpic"/>
+                        <span className="movie-poster-overlay">
+                        <button onClick={(e) => toggleMovieModal(e, movie)}>click me</button>
                         </span>
-                    </section>
                         <div className="recent-movie-details-container">
                             <h5>
                                 {movie.title}
@@ -122,7 +114,7 @@ const toggleMovieModal = (e, movie) => {
                             <ul className="recent-movie-details">
                                 <li>{movie.release_date}</li>
                                 <li>{movie.category}</li>
-                                <li>{movie.length}</li>
+                                <li>2hr 3mins</li>
                             </ul>
                         </div>
                     </div>
@@ -135,67 +127,3 @@ const toggleMovieModal = (e, movie) => {
 }
 
 export {RecentUpload }
-
-
-
-
-const AllMovies = () => {
-
-    const [movies, setMovies] = useState([]);
-    const [isLoading, setIsLoading] = useState(true);
-
-useEffect(() => {
-    const fetchMovies = async () => {
-        const result = await axios(`https://rewind-test.herokuapp.com/movies`)
-        console.log(result.data.message);
-
-        setMovies(result.data.message);
-        setIsLoading(false);
-    }
-    fetchMovies();
-}, [])
-
-const [movieModal, setMovieModal] = useState(false);
-const [selectedMovie, setSelectedMovie] = useState(null);
-
-const toggleMovieModal = (e, movie) => {
-    setMovieModal(prev => !prev);
-    setSelectedMovie(movie);
-    
-}
-
-
-    return  isLoading ? <h1 className="spinner"><CgSpinnerTwo/></h1> : <main>
-        <section className="slider">
-            <h2 id="recent-uploads">
-                All Movies
-            </h2>
-          <section className="recent-upload-container">
-                {movies.map(movie => (
-                    <div key={movie._id} className="recent-upload-content-container" data-aos="fade-up" data-aos-duration="1000">
-                      <section className="movie-poster">
-                        <img src={movie.coverpics_url} alt="poster" />
-                        <span className="movie-poster-overlay" >
-                            <button onClick={(e) => toggleMovieModal(e, movie)}><FiPlay/></button> 
-                        </span>
-                    </section>
-                        <div className="recent-movie-details-container">
-                            <h5>
-                                {movie.title}
-                            </h5>
-                            <ul className="recent-movie-details">
-                                <li>{movie.release_date}</li>
-                                <li>{movie.category}</li>
-                                <li>{movie.length}</li>
-                            </ul>
-                        </div>
-                    </div>
-                ))}
-            <MovieModal movieModal={movieModal} selectedMovie={selectedMovie}  setMovieModal={setMovieModal} />
-         </section>
-    </section> 
-        </main>
-    
-}
-
-export { AllMovies }
