+++ src/components/tabs/tabs.css
@@ -1,3 +1,7 @@
-TabNav {
+#tabnav{
+    background-color: rgba(0, 0, 0, 0.8);
+    color: #fff;
+}
+.holder {
     background-color: rgba(0, 0, 0, 0.8);
 }
