+++ src/components/tabs/tabs.js
@@ -1,12 +1,13 @@
 import React from 'react'
+import "./tabs.css"
 
 const Tab = (props) => {
    
    if(props.isSelected) {
        return (
-           <>
+           <main classNames="holder">
             {props.children}
-           </>
+           </main>
        )
    }
