+++ src/components/tabs/tabscard.js
@@ -1,6 +1,8 @@
 import React, {useEffect, useState} from 'react';
 import { CgSpinnerTwo } from 'react-icons/cg';
 import axios from "axios";
+import {MovieModal} from "../../components";
+import "../cards/cards.css"
 
 const TabsCard = () => {
 
@@ -19,31 +21,40 @@ const TabsCard = () => {
         fetchMovies();
     }, [])
 
+
+    const [movieModal, setMovieModal] = useState(false);
+const [selectedMovie, setSelectedMovie] = useState(null);
+
+const toggleMovieModal = (e, movie) => {
+    setMovieModal(prev => !prev);
+    setSelectedMovie(movie);
+    
+}
+
         return  isLoading ? <h1 className="spinner animate__rotateIn"><CgSpinnerTwo/></h1> : <main
         >
-            <section  className="card-grid">
-            {movies.map(movie => (
-                <section className="movie-container-holder"  key={movie._id}>
-                    <div className="movie-container"  data-aos="fade-up" data-aos-duration="1000">
-                        <section className="movie-poster">
-                            <img src={movie.coverpics_url} alt="poster" />
-                        </section>
-                        <section className="movie-title">
-                            <h4>{movie.title}</h4>
-                        </section>
-                        <section className="movie-container-footer">
-                            <ul>
+                  <section className="recent-upload-container">
+                {movies.map(movie => (
+                    <div key={movie._id} className="recent-upload-content-container">
+                        <img src={movie.coverpics_url} alt="coverpic" className="movie-poster"/>
+                        <span className="movie-poster-overlay">
+                        <button onClick={(e) => toggleMovieModal(e, movie)}>click me</button>
+                        </span>
+                        <div className="recent-movie-details-container">
+                            <h5>
+                                {movie.title}
+                            </h5>
+                            <ul className="recent-movie-details">
                                 <li>{movie.release_date}</li>
-                                <li>Coloured</li>
-                                <li>2hr 9mins</li>
+                                <li>{movie.category}</li>
+                                <li>2hr 3mins</li>
                             </ul>
-                        </section>
-    
+                        </div>
                     </div>
-                </section>
-            ))}
-            </section>
-            
+                ))}
+            <MovieModal movieModal={movieModal} selectedMovie={selectedMovie}  setMovieModal={setMovieModal} />
+         </section>
+
         </main>
         
 }
@@ -69,30 +80,39 @@ const TabsActionCard = () => {
         fetchMovies();
     }, [])
 
+    const [movieModal, setMovieModal] = useState(false);
+const [selectedMovie, setSelectedMovie] = useState(null);
+
+const toggleMovieModal = (e, movie) => {
+    setMovieModal(prev => !prev);
+    setSelectedMovie(movie);
+    
+}
+
+
         return  isLoading ? <h1 className="spinner animate__rotateIn"><CgSpinnerTwo/></h1> : <main
         >
-            <section  className="card-grid">
-            {movies.map(movie => (
-                <section className="movie-container-holder"  key={movie._id}>
-                    <div className="movie-container"  data-aos="fade-up" data-aos-duration="1000">
-                        <section className="movie-poster">
-                            <img src={movie.coverpics_url} alt="poster" />
-                        </section>
-                        <section className="movie-title">
-                            <h4>{movie.title}</h4>
-                        </section>
-                        <section className="movie-container-footer">
-                            <ul>
+             <section className="recent-upload-container">
+                {movies.map(movie => (
+                    <div key={movie._id} className="recent-upload-content-container">
+                        <img src={movie.coverpics_url} alt="coverpic"/>
+                        <span className="movie-poster-overlay">
+                        <button onClick={(e) => toggleMovieModal(e, movie)}>open</button>
+                        </span>
+                        <div className="recent-movie-details-container">
+                            <h5>
+                                {movie.title}
+                            </h5>
+                            <ul className="recent-movie-details">
                                 <li>{movie.release_date}</li>
-                                <li>Coloured</li>
-                                <li>2hr 9mins</li>
+                                <li>{movie.category}</li>
+                                <li>2hr 3mins</li>
                             </ul>
-                        </section>
-    
+                        </div>
                     </div>
-                </section>
-            ))}
-            </section>
+                ))}
+            <MovieModal movieModal={movieModal} selectedMovie={selectedMovie}  setMovieModal={setMovieModal} />
+         </section>
             
         </main>
