+++ src/pages/app/app.css
@@ -106,7 +106,9 @@ border-radius: 0px 0px 4.8px 4.8px;
 .recent-uploads {
     background: #121212 ;
 }
-
+.all-movies {
+    background: #121212 ;
+}
