+++ src/pages/app/landingpage.js
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import axios from "axios";
-import { TabsCard, TabsActionCard, TabNav, Tab, CarouselPage, RecentUpload } from "../../components";
+import { TabsCard, TabsActionCard, TabNav, Tab, CarouselPage, RecentUpload, AllMovies } from "../../components";
 import { Footer, Landingpagenav } from "../../widgets"
 import { Carousel } from "react-bootstrap";
 import { HomeAlone, GodFather, SoundofMusic } from "../../assets";
@@ -72,42 +71,8 @@ const Landingpage = (props) => {
             <section className="recent-uploads">
                 <RecentUpload />
             </section>
-            <section>
-                <TabNav tabs={['Adventure', 'Comedy', 'Horror', 'Action', 'Romance', 'Drama', 'Sci-fi', 'Mystery', 'Animation', 'Nigerian', 'Others']} selected={selected} setSelected={setSelected}>
-                    <Tab isSelected={selected === 'Adventure'}>
-                        <TabsCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Comedy'}>
-                        <p>sdjdnccudjcdcd</p>
-                    </Tab>
-                    <Tab isSelected={selected === 'Horror'}>
-                        <p>ddcodcodcdocdoc</p>
-                    </Tab>
-                    <Tab isSelected={selected === 'Action'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Romance'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Drama'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Sci-fi'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Mystery'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Animation'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Nigerian'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Others'}>
-                        <TabsActionCard />
-                    </Tab>
-                </TabNav>
+            <section className="all-movies">
+                <AllMovies/>
             </section>
             <Footer />
         </>
