+++ src/pages/app/userlandingpage.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import axios from "axios";
-import { TabsCard, TabsActionCard, TabNav, Tab, CarouselPage, RecentUpload } from "../../components";
-import { Footer, Landingpagenav } from "../../widgets"
+import { TabsCard, TabsActionCard, TabNav, Tab, CarouselPage, RecentUpload, AllMovies } from "../../components";
+import { Footer, Userlandingpagenav } from "../../widgets"
 import { Carousel } from "react-bootstrap";
 import { HomeAlone, GodFather, SoundofMusic } from "../../assets";
 import "./app.css";
@@ -30,7 +30,7 @@ const Userlandingpage = (props) => {
 
     return (
         <>
-            <Landingpagenav />
+            <Userlandingpagenav />
 
             <section className="intro-section">
                 <div className="overlay">
@@ -72,43 +72,9 @@ const Userlandingpage = (props) => {
             <section className="recent-uploads">
                 <RecentUpload />
             </section>
-            <section>
-                <TabNav tabs={['Adventure', 'Comedy', 'Horror', 'Action', 'Romance', 'Drama', 'Sci-fi', 'Mystery', 'Animation', 'Nigerian', 'Others']} selected={selected} setSelected={setSelected}>
-                    <Tab isSelected={selected === 'Adventure'}>
-                        <TabsCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Comedy'}>
-                        <p>sdjdnccudjcdcd</p>
-                    </Tab>
-                    <Tab isSelected={selected === 'Horror'}>
-                        <p>ddcodcodcdocdoc</p>
-                    </Tab>
-                    <Tab isSelected={selected === 'Action'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Romance'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Drama'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Sci-fi'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Mystery'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Animation'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Nigerian'}>
-                        <TabsActionCard />
-                    </Tab>
-                    <Tab isSelected={selected === 'Others'}>
-                        <TabsActionCard />
-                    </Tab>
-                </TabNav>
-            </section>
+           <section className="all-movies">
+                <AllMovies/>
+           </section>
             <Footer />
         </>
     )
